requirements.txt
google-api-python-client
google-cloud-vision==0.35.2
Flask==0.12.3
Pillow >=2.0.0
requests
